much needed code clean up
