+++ variant/validators/polypeptide_sequence_variant_base.py
@@ -1,7 +1,6 @@
 """The module for Polypeptide Sequence Variant Validation."""
 from typing import List
 from abc import abstractmethod
-from requests.exceptions import HTTPError
 from .validator import Validator
 from variant.schemas.classification_response_schema import Classification
 from variant.schemas.token_response_schema import GeneMatchToken
@@ -13,10 +12,6 @@ from variant.tokenizers.caches import AminoAcidCache
 from variant.data_sources import SeqRepoAccess, TranscriptMappings
 from ga4gh.vrs import models
 from ga4gh.core import ga4gh_identify
-from ga4gh.vrs.dataproxy import SeqRepoDataProxy
-from ga4gh.vrs.extras.translator import Translator
-import hgvs.parser
-import requests
 import logging
 
 logger = logging.getLogger('variant')
@@ -37,13 +32,8 @@ class PolypeptideSequenceVariantBase(Validator):
         :param TranscriptMappings transcript_mappings: Access to transcript
             mappings
         """
-        self.transcript_mappings = transcript_mappings
-        self.seq_repo_access = seq_repo_access
-        self._gene_matcher = gene_symbol
+        super().__init__(seq_repo_access, transcript_mappings, gene_symbol)
         self._amino_acid_cache = amino_acid_cache
-        self.dp = SeqRepoDataProxy(seq_repo_access.seq_repo_client)
-        self.tlr = Translator(data_proxy=self.dp)
-        self.hgvs_parser = hgvs.parser.Parser()
 
     def validate(self, classification: Classification) \
             -> List[ValidationResult]:
@@ -137,27 +127,6 @@ class PolypeptideSequenceVariantBase(Validator):
             hgvs_expr = hgvs_expr.replace('=', three_letter)
         return hgvs_expr
 
-    def get_allele_from_hgvs(self, hgvs_expr, errors):
-        """Return allele from a given hgvs_expr.
-
-        :param str hgvs_expr: The HGVS string
-        :param list errors: List of errors
-        :return: Allele as a dictionary
-        """
-        allele = None
-        try:
-            allele = self.tlr.translate_from(hgvs_expr, 'hgvs')
-        except HTTPError:
-            errors.append(f"{hgvs_expr} is an invalid HGVS expression.")
-        except KeyError:
-            errors.append("GA4GH Data Proxy unable to translate sequence "
-                          f"identifier: {hgvs_expr}.")
-        except ValueError:
-            errors.append(f"Unable to parse {hgvs_expr} as hgvs variation")
-        else:
-            allele = allele.as_dict()
-        return allele
-
     def get_allele_from_transcript(self, hgvs_expr, t, errors):
         """Return allele from a given transcript.
 
@@ -203,7 +172,7 @@ class PolypeptideSequenceVariantBase(Validator):
                             'transcript_token': t
                         }
                 else:
-                    refseq = ([a for a in self.seq_repo_access.aliases(t)
+                    refseq = ([a for a in self.seqrepo_access.aliases(t)
                                if a.startswith('refseq:NP_')] or [None])[0]
 
                     if not refseq:
@@ -230,7 +199,7 @@ class PolypeptideSequenceVariantBase(Validator):
                             allele['state']['sequence'])
 
                 ref_protein = \
-                    self.seq_repo_access.sequence_at_position(t, s.position)
+                    self.seqrepo_access.sequence_at_position(t, s.position)
 
                 if not errors:
                     if ref_protein and len(ref_protein) == 1 \
@@ -259,159 +228,6 @@ class PolypeptideSequenceVariantBase(Validator):
         self.add_mane_transcript(classification, results, gene_tokens,
                                  mane_transcripts_dict)
 
-    def add_mane_transcript(self, classification, results, gene_tokens,
-                            mane_transcripts_dict):
-        """Add MANE transcript validation result objects to a list of results.
-
-        :param Classification classification: A classification for a list of
-            tokens
-        :param list results: A list to store validation result objects
-        :param list gene_tokens: List of GeneMatchTokens
-        :param dict mane_transcripts_dict: Possible MANE select transcripts
-            with classification and Ensembl transcript
-        """
-        mane_transcripts = list()
-        found_mane_transcripts = list()
-        replace_old_keys = list()
-        for hgvs_expr in mane_transcripts_dict.keys():
-            if '(' in hgvs_expr and ')' in hgvs_expr:
-                # ClinGen Allele Registry doesn't like () in the query
-                if mane_transcripts_dict[hgvs_expr][
-                    'classification_token'].token_type ==\
-                        'PolypeptideTruncation':
-                    old = hgvs_expr
-                    hgvs_expr = hgvs_expr.replace('(', '')
-                    hgvs_expr = hgvs_expr.replace(')', '')
-                    old_and_new = old, hgvs_expr
-                    if old_and_new not in replace_old_keys:
-                        replace_old_keys.append(old_and_new)
-            t = self.get_mane_transcript(hgvs_expr)
-            if t and t not in found_mane_transcripts:
-                found_mane_transcripts.append(t)
-                mane_transcripts.append((hgvs_expr, t))
-
-        for tup in replace_old_keys:
-            mane_transcripts_dict[tup[1]] = mane_transcripts_dict[tup[0]]
-            del mane_transcripts_dict[tup[0]]
-
-        errors = list()
-        allele = {}
-        seq_location = None
-
-        if len(mane_transcripts) == 0:
-            logger.warning("No MANE Select transcript found for "
-                           f"{mane_transcripts_dict.keys()}")
-            return
-
-        for el in mane_transcripts:
-            hgvs_expr = el[0]
-            transcript = el[1]
-
-            try:
-                seq_location = self.tlr.translate_from(transcript, 'hgvs')
-            except HTTPError:
-                errors.append(f"{mane_transcripts[0]} is an invalid HGVS "
-                              f"expression.")
-
-            if not errors:
-                allele = seq_location.as_dict()
-                if len(allele['state']['sequence']) == 3:
-                    allele['state']['sequence'] = \
-                        self._amino_acid_cache.convert_three_to_one(
-                            allele['state']['sequence'])
-
-                results.append(self.get_validation_result(
-                    classification, True, 1, allele,
-                    self.human_description(
-                        mane_transcripts_dict[hgvs_expr][
-                            'transcript_token'],
-                        mane_transcripts_dict[hgvs_expr][
-                            'classification_token']
-                    ),
-                    self.concise_description(
-                        mane_transcripts_dict[hgvs_expr][
-                            'transcript_token'],
-                        mane_transcripts_dict[hgvs_expr][
-                            'classification_token']
-                    ), errors, gene_tokens, transcript
-                ))
-            else:
-                results.append(self.get_validation_result(
-                    classification, False, 1, allele,
-                    self.human_description(
-                        mane_transcripts_dict[hgvs_expr]['transcript_token'],
-                        mane_transcripts_dict[hgvs_expr][
-                            'classification_token']
-                    ),
-                    self.concise_description(
-                        mane_transcripts_dict[hgvs_expr]['transcript_token'],
-                        mane_transcripts_dict[hgvs_expr][
-                            'classification_token']
-                    ), errors, gene_tokens, transcript
-                ))
-
-    def get_mane_transcript(self, hgvs_expr):
-        """Return MANE Select Transcript from ClinGene Allele Registry.
-
-        :param str hgvs_expr: The HGVS expression to query
-        :return: The HGVS RefSeq MANE Select Transcript represented as a string
-        """
-        request = requests.get(
-            f"https://reg.genome.network/allele?hgvs={hgvs_expr}")
-        if request.status_code == 200:
-            resp = request.json()
-            if 'transcriptAlleles' in resp.keys():
-                mane_transcript = None
-                for t in resp['transcriptAlleles']:
-                    if 'MANE' in t.keys():
-                        mane_transcript = t['MANE']
-                        break
-                if mane_transcript:
-                    if mane_transcript['maneStatus'] == 'MANE Select':
-                        result = mane_transcript['protein']['RefSeq']['hgvs']
-                        return result
-            else:
-                if 'aminoAcidAlleles' in resp.keys() and len(resp['aminoAcidAlleles']) > 0:  # noqa: E501
-                    amino_acid_allele = resp['aminoAcidAlleles'][0]
-                    if 'hgvsMatchingTranscriptVariant' in amino_acid_allele.keys():  # noqa: E501
-                        if len(amino_acid_allele['hgvsMatchingTranscriptVariant']) > 0:  # noqa: E501
-                            for t in amino_acid_allele['hgvsMatchingTranscriptVariant']:  # noqa: E501
-                                if '>' in t and '[' not in t:
-                                    # Temp condition since variant norm
-                                    # cannot handle multiple possible variants
-                                    return self.get_mane_transcript(t)
-        return None
-
-    def get_validation_result(self, classification, is_valid, confidence_score,
-                              allele, human_description, concise_description,
-                              errors, gene_tokens,
-                              mane_transcript=None) -> ValidationResult:
-        """Return a validation result object.
-
-        :param Classification classification: The classification for tokens
-        :param boolean is_valid: Whether or not the classification is valid
-        :param int confidence_score: The classification confidence score
-        :param dict allele: A VRS Allele object
-        :param str human_description: A human description describing the
-            variant
-        :param str concise_description: The identified variant
-        :param list errors: A list of errors for the classification
-        :param list gene_tokens: List of GeneMatchTokens
-        :param str mane_transcript: MANE transcript
-        :return: A validation result
-        """
-        return ValidationResult(
-            classification=classification,
-            is_valid=is_valid,
-            confidence_score=confidence_score,
-            allele=allele,
-            human_description=human_description,
-            concise_description=concise_description,
-            errors=errors,
-            gene_tokens=gene_tokens,
-            mane_transcript=mane_transcript
-        )
-
     def get_gene_tokens(self, classification) -> List[GeneMatchToken]:
         """Return gene tokens for a classification.
 
@@ -432,7 +248,7 @@ class PolypeptideSequenceVariantBase(Validator):
             if ':' in refseq:
                 refseq = refseq.split(':')[0]
 
-            res = self.seq_repo_access.aliases(refseq)
+            res = self.seqrepo_access.aliases(refseq)
             aliases = [a.split('ensembl:')[1] for a
                        in res if a.startswith('ensembl')]
 
@@ -459,22 +275,6 @@ class PolypeptideSequenceVariantBase(Validator):
                         gene_tokens.append(self._gene_matcher.match(
                             gene_symbol))
         return gene_tokens
-
-    @abstractmethod
-    def is_token_instance(self, t) -> bool:
-        """Check to see if token is instance of a token type."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def variant_name(self):
-        """Return the variant name"""
-        raise NotImplementedError
-
-    @abstractmethod
-    def validates_classification_type(self, classification_type) \
-            -> bool:
-        """Return the classification type"""
-        raise NotImplementedError
 
     def concise_description(self, transcript, token) -> str:
         """Return a description of the identified variant."""
